+++ mycodo/devices/camera.py
@@ -12,13 +12,11 @@ from mycodo.databases.models import OutputChannel
 from mycodo.databases.utils import session_scope
 from mycodo.mycodo_client import DaemonControl
 from mycodo.utils.database import db_retrieve_table_daemon
-from mycodo.utils.logging_utils import set_log_level
 from mycodo.utils.system_pi import assure_path_exists
 from mycodo.utils.system_pi import cmd_output
 from mycodo.utils.system_pi import set_user_grp
 
 logger = logging.getLogger(__name__)
-logger.setLevel(set_log_level(logging))
 
 
 #