+++ mycodo/mycodo_daemon.py
@@ -36,13 +36,13 @@ import threading
 import time
 import timeit
 import traceback
+from logging import handlers
 
-from daemonize import Daemonize
 from Pyro5.api import Proxy, expose, serve
 
-from mycodo.config import (DAEMON_LOG_FILE, DAEMON_PID_FILE, DOCKER_CONTAINER,
-                           MYCODO_DB_PATH, MYCODO_VERSION, STATS_CSV,
-                           STATS_INTERVAL, UPGRADE_CHECK_INTERVAL)
+from mycodo.config import (DAEMON_LOG_FILE, DOCKER_CONTAINER, MYCODO_DB_PATH,
+                           MYCODO_VERSION, STATS_CSV, STATS_INTERVAL,
+                           UPGRADE_CHECK_INTERVAL)
 from mycodo.controllers.controller_conditional import ConditionalController
 from mycodo.controllers.controller_function import FunctionController
 from mycodo.controllers.controller_input import InputController
@@ -65,24 +65,30 @@ from mycodo.utils.stats import (add_update_csv, recreate_stat_file,
 from mycodo.utils.tools import generate_output_usage_report, next_schedule
 
 
+formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(name)s - %(message)s')
+logHandler = logging.FileHandler(DAEMON_LOG_FILE)
+logHandler.setLevel(logging.DEBUG)
+logHandler.setFormatter(formatter)
+
+logger = logging.getLogger('mycodo')
+logger.setLevel(logging.DEBUG)
+logger.addHandler(logHandler)
+
+
 class DaemonController:
     """Mycodo daemon."""
-    def __init__(self, debug):
-        self.logger = logging.getLogger('mycodo.daemon')
-        if not debug:
-            self.logger.setLevel(logging.INFO)
+    def __init__(self):
+        self.logger = logger
 
         self.logger.info(f"Mycodo daemon v{MYCODO_VERSION} starting")
 
         if DOCKER_CONTAINER:
             self.logger.info("Detected running inside a Docker continaer")
 
-        self.log_level_debug = debug
         self.startup_timer = timeit.default_timer()
         self.startup_time = None
         self.daemon_run = True
         self.terminated = False
-        self.debug = debug
 
         # Actions
         self.actions = {}
@@ -136,7 +142,7 @@ class DaemonController:
         self.load_actions()
 
         try:
-            self.start_all_controllers(self.debug)
+            self.start_all_controllers()
         except Exception:
             self.logger.exception("Could not start all controllers")
 
@@ -850,7 +856,7 @@ class DaemonController:
         self.actions = parse_action_information()
 
 
-    def start_all_controllers(self, debug):
+    def start_all_controllers(self):
         """
         Start all activated controllers
 
@@ -1107,7 +1113,7 @@ class DaemonController:
 
         # Send stats
         try:
-            send_anonymous_stats(self.start_time, self.log_level_debug)
+            send_anonymous_stats(self.start_time)
         except Exception:
             self.logger.exception("Sending statistics")
 
@@ -1313,14 +1319,10 @@ class PyroDaemon(threading.Thread):
     user) to communicate with the daemon (mycodo_daemon.py running with root privileges).
 
     """
-    def __init__(self, mycodo, debug):
+    def __init__(self, mycodo):
         threading.Thread.__init__(self)
 
         self.logger = logging.getLogger('mycodo.pyro_daemon')
-        if debug:
-            self.logger.setLevel(logging.DEBUG)
-        else:
-            self.logger.setLevel(logging.WARNING)
         self.mycodo = mycodo
 
     def run(self):
@@ -1338,14 +1340,10 @@ class PyroMonitor(threading.Thread):
     Monitor whether the Pyro5 server (and daemon) is active or not
 
     """
-    def __init__(self, debug):
+    def __init__(self):
         threading.Thread.__init__(self)
 
         self.logger = logging.getLogger('mycodo.pyro_monitor')
-        if debug:
-            self.logger.setLevel(logging.DEBUG)
-        else:
-            self.logger.setLevel(logging.WARNING)
         self.timer_sec = 1800
 
     def run(self):
@@ -1372,24 +1370,20 @@ class MycodoDaemon:
     """
     Handle starting the components of the Mycodo Daemon
     """
-    def __init__(self, mycodo, debug):
+    def __init__(self, mycodo):
         self.logger = logging.getLogger('mycodo.daemon')
-        if not debug:
-            self.logger.setLevel(logging.INFO)
-        self.debug = debug
         self.mycodo = mycodo
 
     def start_daemon(self):
         """Start communication and daemon threads."""
         try:
-            pd = PyroDaemon(self.mycodo, self.debug)
+            pd = PyroDaemon(self.mycodo)
             pd.daemon = True
             pd.start()
 
-            if self.debug:
-                pm = PyroMonitor(self.debug)
-                pm.daemon = True
-                pm.start()
+            # pm = PyroMonitor()
+            # pm.daemon = True
+            # pm.start()
 
             self.mycodo.run()  # Start daemon thread that manages controllers
         except Exception:
@@ -1410,25 +1404,8 @@ if __name__ == '__main__':
     if not os.geteuid() == 0:
         sys.exit("Script must be executed as root")
 
-    # Check if lock file already exists
-    # if os.path.isfile(DAEMON_PID_FILE):
-    #     sys.exit(
-    #         "Daemon PID file present. Ensure the daemon isn't already "
-    #         f"running and delete {DAEMON_PID_FILE}")
-
     # Parse commandline arguments
     args = parse_args()
-
-    # Set up logger
-    logger = logging.getLogger('mycodo')
-    logger.setLevel(logging.DEBUG)
-    logger.propagate = False
-    fh = logging.FileHandler(DAEMON_LOG_FILE, 'a')
-    fh.setLevel(logging.DEBUG)
-    formatter = logging.Formatter("%(asctime)s - %(levelname)s - %(name)s - %(message)s")
-    fh.setFormatter(formatter)
-    logger.addHandler(fh)
-    keep_fds = [fh.stream.fileno()]
 
     debug = False
     misc = db_retrieve_table_daemon(Misc, entry='first')
@@ -1437,16 +1414,13 @@ if __name__ == '__main__':
     if args.debug:
         debug = args.debug
 
-    daemon_controller = DaemonController(debug)
-    mycodo_daemon = MycodoDaemon(daemon_controller, debug)
-
-    if DOCKER_CONTAINER:
-        mycodo_daemon.start_daemon()
+    if debug:
+        log_level = logging.DEBUG
     else:
-        # Set up daemon and start it
-        daemon = Daemonize(
-            app="mycodo_daemon",
-            pid=DAEMON_PID_FILE,
-            action=mycodo_daemon.start_daemon,
-            keep_fds=keep_fds)
-        daemon.start()
+        log_level = logging.INFO
+
+    logger.setLevel(log_level)
+
+    daemon_controller = DaemonController()
+    mycodo_daemon = MycodoDaemon(daemon_controller)
+    mycodo_daemon.start_daemon()