+++ mycodo/utils/code_verification.py
@@ -5,13 +5,11 @@ from flask import Markup
 
 from mycodo.config import INSTALL_DIRECTORY
 from mycodo.config import PATH_PYTHON_CODE_USER
-from mycodo.utils.logging_utils import set_log_level
 from mycodo.utils.system_pi import assure_path_exists
 from mycodo.utils.system_pi import cmd_output
 from mycodo.utils.system_pi import set_user_grp
 
 logger = logging.getLogger(__name__)
-logger.setLevel(set_log_level(logging))
 
 
 def create_python_file(python_code_run, filename):