+++ mycodo/utils/lockfile.py
@@ -5,10 +5,7 @@ import time
 
 import filelock
 
-from mycodo.utils.logging_utils import set_log_level
-
 logger = logging.getLogger("mycodo.lockfile")
-logger.setLevel(set_log_level(logging))
 
 
 class LockFile: