+++ mycodo/utils/send_data.py
@@ -9,10 +9,7 @@ from email.mime.base import MIMEBase
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 
-from mycodo.utils.logging_utils import set_log_level
-
 logger = logging.getLogger("mycodo.notification")
-logger.setLevel(set_log_level(logging))
 
 
 #